+++ lang/english.php
@@ -41,3 +41,7 @@ $strings['badResponseErrors'] = 'Bad Response. Errors: ';
 $strings['mixedContent'] = 'Mixed Active Content is not allowed. HTTPS address for ONLYOFFICE Docs is required.';
 $strings['badHealthcheckStatus'] = 'Bad healthcheck status';
 $strings['notSupportedVersion'] = 'Not supported version';
+$strings['UserVoiceTitle'] = 'Welcome to ONLYOFFICE Docs!';
+$strings['UserVoiceDescription'] = 'Collaborate on documents, spreadsheets, presentations, and forms within the Chamilo structure using ONLYOFFICE Docs.';
+$strings['LearnMore'] = 'Learn more';
+$strings['SuggestFeature'] = 'Suggest a feature';