+++ lib/onlyofficeSettingsFormBuilder.php
@@ -14,7 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 require_once __DIR__.'/../../../main/inc/global.inc.php';
 
 class OnlyofficeSettingsFormBuilder
@@ -74,6 +73,9 @@ class OnlyofficeSettingsFormBuilder
             'banner_main_text' => $plugin->get_lang('DocsCloudBannerMain'),
             'banner_button_text' => $plugin->get_lang('DocsCloudBannerButton'),
         ]);
+        $userVoiceTemplate = self::buildTemplate('user_voice');
+        $userVoice = $plugin_info['settings_form']->createElement('html', $userVoiceTemplate);
+        $plugin_info['settings_form']->insertElementBefore($userVoice, 'document_server_url');
         $plugin_info['settings_form']->insertElementBefore($connectDemoCheckbox, 'submit_button');
         $demoServerMessage = $plugin_info['settings_form']->createElement('html', $demoServerMessageHtml);
         $plugin_info['settings_form']->insertElementBefore($demoServerMessage, 'submit_button');